Electric Bill v0.1
I plan to continue developing this one someday to be much more complex. I'm sure there are plenty of apps like this already but id like to someday say ive created a program tailored more to my liking and of course if i complete it, its free lol. Think i might need a lot more practice building more complex programs with more elegance...there is a lot going on here and i do have to do so much more to it.