Update Electric Bill v0.1
Had to change the notes that did not reflect the original idea of the on peak/off peak/super off peak hours. Also needed to make it more dynamic in order for others to input their own electrical rate. I will now prompt for what you rate is at the beginning.