Update Scrambled v.1.0
This app is the my first version of what I'm calling "Scrambled" for now. This is meant to help create passwords that are memorable. A series of very memorable phrases are meant to be used in order to allow for the safe keeping of passwords in mind as oppose to paper. The options are limited to 8, 12, 16 and 20. I hope to work on making it more user friendly and work out the error codes so to make it more app friendly of course. My path through learning python is very fun and i hope that one day i can look back and see how simply these programs were and laugh haha.